Split syntax and hir visitors
